move helper methods to their own file as functions
